Added 4 Lua Functions & 1 Callback (#12)
* Updated Additional Includes to be relative to the project

This will make it easier for others as before hand we would have to
change the the additional includes to match our paths. Where as this way
it is relative to the project

* Added Missing ScriptHookV.lib

Not sure why this isn't in the files. Also should mention the Commit
Before updated the additional library location too to be relative tot he
project.

* Revert "Added Missing ScriptHookV.lib"

This reverts commit 007a1e2444d0c175a87381808bc51ab9eeb48591.

* Revert "Revert "Added Missing ScriptHookV.lib""

This reverts commit 668e632b7aaa32895986f248c1f8adcfbdd06b32.

* Merge remote-tracking branch 'refs/remotes/FiveMP/master'

Second merge to fix the conflicts as to resolve it i had to remove the
files.

* LUA: Added Callback OnPlayerDisconnect(playerid)

Called when a player disconnects and also when they lose connection.

* Added Give/Remove Weapon & Ammo functions

GivePlayerWeapon(playerid, weaponid, ammo); //ammo -1 is
Infin
RemovePlayerWeapon(weaponid);
GivePlayerAmmo(playerid, weaponid,
ammo);
RemovePlayerAmmo(playerid, weaponid, ammo);

* Added and Example of GivePlayerWeapon()

* Added example OnPlayerDisconnect